Made a couple fixes to 01_setup_workspace which fixed a bug that didn't allow you to copy workspaces that included a rosetta directory. Also, made sure that setup.py isntalls everything in standard_params.

+++ pull_into_place/commands/01_setup_workspace.py
@@ -85,8 +85,8 @@ the symlink called 'rosetta' in the workspace directory."""
                 if not os.path.exists(path):
                     message.append('    ' + path)
             raise ValueError('\n'.join(message))
-
-        os.symlink(rosetta_dir, workspace.rosetta_dir)
+	if not os.path.exists(workspace.rosetta_dir):
+	    os.symlink(rosetta_dir, workspace.rosetta_dir)
 
 
 class ProjectParams(Installer):
